+++ tests/testthat/test.geom_dotsinterval.R
@@ -294,7 +294,6 @@ test_that("dot order is correct", {
   skip_if_no_vdiffr()
 
   p = data.frame(x = qnorm(ppoints(50)), g = c("a", "b")) %>%
-    arrange(g) %>%
     ggplot(aes(x = x, fill = stat(x < 0), color = g, group = NA)) +
     scale_fill_brewer(palette = "Set1") +
     scale_color_brewer(palette = "Paired")
@@ -307,7 +306,7 @@ test_that("dot order is correct", {
 
   vdiffr::expect_doppelganger("bin dot order, kept",
     p +
-      geom_dots(layout = "bin", size = 5, keep_order = TRUE) +
+      geom_dots(aes(order = g), layout = "bin", size = 5) +
       geom_vline(xintercept = 0)
   )